Rename QT_VERSION_MAJOR to QT_MAJOR_VERSION
This is the common name in KDE and set eg. by the CI system.

This change also makes setting the Qt Major Version explicitely
required. It does no longer look for Qt5 and Qt6 and uses the one it
finds. If no version is set it always defaults to Qt5 and fails if Qt5
is not there (even if Qt6 is)